requirements.txt
kaggle
tensorflow
scikit-learn
matplotlib
seaborn
pyyaml
python-box